+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/case-viewer/case-event-trigger/case-event-trigger.component.ts
@@ -2,7 +2,7 @@ import { Component, NgZone, OnDestroy, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Observable, of } from 'rxjs';
 import { Subscription } from 'rxjs/Subscription';
-import { Activity, CaseEventData, CaseEventTrigger, CaseField, CaseView, DisplayMode } from '../../../domain';
+import { Activity, CaseEventData, CaseEventTrigger, CaseView, DisplayMode } from '../../../domain';
 import { CaseReferencePipe } from '../../../pipes';
 import { ActivityPollingService, AlertService, EventStatusService, FieldsUtils } from '../../../services';
 import { CaseNotifier, CasesService } from '../../case-editor';
@@ -72,21 +72,11 @@ export class CaseEventTriggerComponent implements OnInit, OnDestroy {
 
   public validate(): (sanitizedEditForm: CaseEventData, pageId: string) => Observable<object> {
     return (sanitizedEditForm: CaseEventData, pageId: string) => {
-      // Bypass validation if the CaseEventData data object contains a FlagLauncher field; this field type cannot be
-      // validated like regular fields. Need to match this field id against that of the defined FlagLauncher CaseField
-      // (if it exists on any CaseTab)
-      const flagLauncherCaseFields: CaseField[] = [];
-      if (this.caseDetails.tabs) {
-        for (const tab of this.caseDetails.tabs) {
-          if (tab.fields) {
-            flagLauncherCaseFields.push(...tab.fields.filter(caseField => FieldsUtils.isFlagLauncherCaseField(caseField)));
-          }
-        }
-      }
-
-      const isThereAnyFlagLauncherField = flagLauncherCaseFields.length > 0;
-
-      return isThereAnyFlagLauncherField ? of(null) : this.casesService.validateCase(this.caseDetails.case_type.id, sanitizedEditForm, pageId);
+      // Bypass validation if the eventTrigger case fields contain a FlagLauncher field; this field type cannot be
+      // validated, unlike regular fields
+      return this.eventTrigger?.case_fields?.some((caseField) => FieldsUtils.isCaseFieldOfType(caseField, ['FlagLauncher']))
+        ? of(null)
+        : this.casesService.validateCase(this.caseDetails.case_type.id, sanitizedEditForm, pageId);
     };
   }